BUGFIX : Plus de mort en touchant le pic Le pic ne fait plus mourir si l'on n'est pas proche de la surface du pic (marcher depuis la gauche fait juste bloquer, par exemple)

+++ ressources.h
@@ -77,7 +77,7 @@ const sf::Color couleur{ sf::Color(128, 128, 128) };
 	constexpr int tempsParImage{ 1000000 / 60 };				//Temps en microsecondes
 namespace utilitaire {
 	constexpr unsigned int limiteFramerate{ 60 };
-	constexpr int deplacement{ 3 };
+	constexpr float deplacement{ 3.f };
 }
 
 bool verifTouches(const ensembleTouches& pTouches);
